Re-organise tests
